--- mpa/modules/datasets/pipelines/torchvision2mmseg.py
@@ -1,295 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-from mmseg.datasets import PIPELINES
-from mmcv.utils import build_from_cfg
-
-import mmcv
-import numpy as np
-# import cv2 as cv
-from torchvision import transforms as T
-# import torchvision.transforms.functional as F
-from mmseg.datasets.pipelines.formating import ImageToTensor, to_tensor
-# from mmdet.datasets.pipelines.transforms import Normalize
-from PIL import Image, ImageFilter
-
-from mmcv.parallel import DataContainer as DC
-
-
-
-@PIPELINES.register_module(force=True)
-class Normalize(object):
-    """Normalize the image.
-
-    Added key is "img_norm_cfg".
-
-    Args:
-        mean (sequence): Mean values of 3 channels.
-        std (sequence): Std values of 3 channels.
-        to_rgb (bool): Whether to convert the image from BGR to RGB,
-            default is true.
-    """
-
-    def __init__(self, mean, std, to_rgb=True):
-        self.mean = np.array(mean, dtype=np.float32)
-        self.std = np.array(std, dtype=np.float32)
-        self.to_rgb = to_rgb
-
-    def __call__(self, results):
-        """Call function to normalize images.
-
-        Args:
-            results (dict): Result dict from loading pipeline.
-
-        Returns:
-            dict: Normalized results, 'img_norm_cfg' key is added into
-                result dict.
-        """
-
-        for target in ['img', 'ul_w_img', 'aux_img']:
-            if target in results:
-                results[target] = mmcv.imnormalize(results[target], self.mean, self.std, self.to_rgb)
-        results['img_norm_cfg'] = dict(mean=self.mean, std=self.std, to_rgb=self.to_rgb)
-
-        return results
-
-    def __repr__(self):
-        repr_str = self.__class__.__name__
-        repr_str += f'(mean={self.mean}, std={self.std}, to_rgb=' \
-                    f'{self.to_rgb})'
-        return repr_str
-
-
-@PIPELINES.register_module(force=True)
-class DefaultFormatBundle(object):
-    """Default formatting bundle.
-
-    It simplifies the pipeline of formatting common fields, including "img"
-    and "gt_semantic_seg". These fields are formatted as follows.
-
-    - img: (1)transpose, (2)to tensor, (3)to DataContainer (stack=True)
-    - gt_semantic_seg: (1)unsqueeze dim-0 (2)to tensor,
-                       (3)to DataContainer (stack=True)
-    """
-
-    def __call__(self, results):
-        """Call function to transform and format common fields in results.
-
-        Args:
-            results (dict): Result dict contains the data to convert.
-
-        Returns:
-            dict: The result dict contains the data that is formatted with
-                default bundle.
-        """
-        for target in ['img', 'ul_w_img', 'aux_img']:
-            if target not in results:
-                continue
-
-            img = results[target]
-            if len(img.shape) < 3:
-                img = np.expand_dims(img, -1)
-            img = np.ascontiguousarray(img.transpose(2, 0, 1)).astype(np.float32)
-
-            results[target] = DC(to_tensor(img), stack=True)
-
-        for trg_name in ['gt_semantic_seg', 'gt_class_borders', 'pixel_weights']:
-            if trg_name not in results:
-                continue
-
-            out_type = np.float32 if trg_name == 'pixel_weights' else np.int64
-            results[trg_name] = DC(
-                to_tensor(results[trg_name][None, ...].astype(out_type)),
-                stack=True
-            )
-
-        return results
-
-    def __repr__(self):
-        return self.__class__.__name__
-
-
-@PIPELINES.register_module()
-class ColorJitter(T.ColorJitter):
-    """MMDet adapter"""
-
-    def __init__(self, key_maps=[('img', 'img')], **kwargs):
-        super().__init__(**kwargs)
-        self.key_maps = key_maps
-
-    def forward(self, inputs):
-        outputs = inputs.copy()
-        for key_map in self.key_maps:
-            outputs[key_map[0]] = super().forward(inputs[key_map[1]])
-        return outputs
-
-
-@PIPELINES.register_module()
-class RandomGrayscale(T.RandomGrayscale):
-    """MMDet adapter"""
-
-    def __init__(self, key_maps=[('img', 'img')], **kwargs):
-        super().__init__(**kwargs)
-        self.key_maps = key_maps
-
-    def forward(self, inputs):
-        outputs = inputs.copy()
-        for key_map in self.key_maps:
-            outputs[key_map[0]] = super().forward(inputs[key_map[1]])
-        return outputs
-
-
-@PIPELINES.register_module()
-class RandomErasing(T.RandomErasing):
-    """MMDet adapter"""
-
-    def __init__(self, key_maps=[('img', 'img')], **kwargs):
-        super().__init__(**kwargs)
-        self.key_maps = key_maps
-
-    def forward(self, inputs):
-        outputs = inputs.copy()
-        for key_map in self.key_maps:
-            outputs[key_map[0]] = super().forward(inputs[key_map[1]])
-        return outputs
-
-
-@PIPELINES.register_module()
-class RandomGaussianBlur(object):
-    """Gaussian blur augmentation in SimCLR https://arxiv.org/abs/2002.05709."""
-
-    def __init__(self, sigma_min, sigma_max, key_maps=[('img', 'img')]):
-        self.sigma_min = sigma_min
-        self.sigma_max = sigma_max
-        self.key_maps = key_maps
-
-    def __call__(self, inputs):
-        outputs = inputs.copy()
-        sigma = np.random.uniform(self.sigma_min, self.sigma_max)
-        # ksize = 2*int(np.ceil(2.0*sigma)) + 1
-        for key_map in self.key_maps:
-            img = inputs[key_map[0]]
-            img = img.filter(ImageFilter.GaussianBlur(radius=sigma))
-            # img = cv.GaussianBlur(img, ksize=(0,0), sigmaX=sigma)
-            # img = F.gaussian_blur(img, ksize, [sigma, sigma])
-            outputs[key_map[1]] = img
-        return outputs
-
-    def __repr__(self):
-        repr_str = self.__class__.__name__
-        return repr_str
-
-
-@PIPELINES.register_module()
-class RandomApply(T.RandomApply):
-    """MMDet adapter"""
-
-    def __init__(self, transform_cfgs, p=0.5):
-        transforms = []
-        for transform_cfg in transform_cfgs:
-            transforms.append(build_from_cfg(transform_cfg, PIPELINES))
-        super().__init__(transforms, p=p)
-
-
-@PIPELINES.register_module()
-class NDArrayToTensor(ImageToTensor):
-    """MMDet adapter"""
-
-    def __call__(self, results):
-        for key in self.keys:
-            img = results[key]
-            if len(img.shape) < 3:
-                img = np.expand_dims(img, -1)
-            img = np.ascontiguousarray(img.transpose(2, 0, 1))
-            results[key] = to_tensor(img)
-        return results
-
-
-# @PIPELINES.register_module()
-# class NormalizeTensor(Normalize):
-#     """MMDet adapter"""
-#
-#     def __call__(self, results):
-#         for key in results.get('img_fields', ['img']):
-#             img = results[key]
-#             img = F.normalize(img.float(), self.mean, self.std)
-#             if self.to_rgb:
-#                 img = img[[2, 1, 0]]
-#             results[key] = img
-#         results['img_norm_cfg'] = dict(
-#             mean=self.mean, std=self.std, to_rgb=self.to_rgb)
-#         return results
-#
-#
-# @PIPELINES.register_module()
-# class PadTensor(object):
-#     def __init__(self, size_divisor=32):
-#         self.size_divisor = float(size_divisor)
-#
-#     def __call__(self, results):
-#         for key in results.get('img_fields', ['img']):
-#             img = results[key]
-#             h = img.shape[1]
-#             w = img.shape[2]
-#             H = int(np.ceil(h/self.size_divisor)*self.size_divisor)
-#             W = int(np.ceil(w/self.size_divisor)*self.size_divisor)
-#             padding = (0, 0, W-w, H-h)
-#             # print(padding)
-#             img = F.pad(img, padding)
-#             results[key] = img
-#             results['pad_shape'] = img.shape
-#         results['pad_size_divisor'] = self.size_divisor
-#         return results
-
-
-@PIPELINES.register_module()
-class NDArrayToPILImage(object):
-    def __init__(self, keys=['img']):
-        self.keys = keys
-
-    def __call__(self, results):
-        for key in self.keys:
-            img = results[key]
-            img = Image.fromarray(img, mode='RGB')
-            results[key] = img
-        return results
-
-    def __repr__(self):
-        repr_str = self.__class__.__name__
-        return repr_str
-
-
-@PIPELINES.register_module()
-class PILImageToNDArray(object):
-    def __init__(self, keys=['img']):
-        self.keys = keys
-
-    def __call__(self, results):
-        for key in self.keys:
-            img = results[key]
-            img = np.asarray(img)
-            results[key] = img
-        return results
-
-    def __repr__(self):
-        repr_str = self.__class__.__name__
-        return repr_str
-
-
-@PIPELINES.register_module()
-class BranchImage(object):
-    def __init__(self, key_map={}):
-        self.key_map = key_map
-
-    def __call__(self, results):
-        for k1, k2 in self.key_map.items():
-            if k1 in results:
-                results[k2] = results[k1]
-            if k1 in results['img_fields']:
-                results['img_fields'].append(k2)
-        return results
-
-    def __repr__(self):
-        repr_str = self.__class__.__name__
-        return repr_str
