+++ mpa/seg/__init__.py
@@ -9,9 +9,9 @@ from . import inferrer
 from . import stage
 from . import trainer
 
-import mpa.modules.datasets.pipelines.torchvision2mmseg
+import mpa.modules.datasets.pipelines.seg_custom_pipelines
 
-import mpa.modules.datasets.pseudo_sseg_dataset
+import mpa.modules.datasets.seg_unlabeled_dataset
 import mpa.modules.datasets.seg_incr_cityscapes_dataset
 import mpa.modules.datasets.seg_incr_voc_dataset
 import mpa.modules.datasets.seg_task_adapt_dataset
@@ -24,5 +24,4 @@ import mpa.modules.models.heads.custom_ocr_head
 import mpa.modules.models.losses.am_softmax_loss_with_ignore
 import mpa.modules.models.losses.cross_entropy_loss_with_ignore
 import mpa.modules.models.losses.recall_loss
-import mpa.modules.models.losses.mse_loss
