+++ recipes/stages/_base_/data/kvasir_seg_semi.py
@@ -1,7 +1,5 @@
 _base_ = [
     './data_seg.py',
-    # './pipelines/incr_seg.py',
-    './pipelines/semi_seg.py'
 ]
 
 __dataset_type = 'KvasirDataset'
